# [wasm] WebSocket improve send/receive buffers