--- a/src/libraries/Common/tests/System/Net/Prerequisites/NetCoreServer/Handlers/EchoWebSocketHandler.cs
+++ b/src/libraries/Common/tests/System/Net/Prerequisites/NetCoreServer/Handlers/EchoWebSocketHandler.cs
@@ -26,6 +26,10 @@ namespace NetCoreServer
             {
                 Thread.Sleep(10000);
             }
+            else if (context.Request.QueryString.HasValue && context.Request.QueryString.Value.Contains("delay20sec"))
+            {
+                Thread.Sleep(20000);
+            }
 
             try
             {