--- a/src/libraries/System.Net.WebSockets.Client/src/System/Net/WebSockets/WebSocketHandle.Browser.cs
+++ b/src/libraries/System.Net.WebSockets.Client/src/System/Net/WebSockets/WebSocketHandle.Browser.cs
@@ -8,7 +8,6 @@ namespace System.Net.WebSockets
 {
     internal sealed class WebSocketHandle
     {
-        private readonly CancellationTokenSource _abortSource = new CancellationTokenSource();
         private WebSocketState _state = WebSocketState.Connecting;
 
         public WebSocket? WebSocket { get; private set; }
@@ -24,53 +23,17 @@ namespace System.Net.WebSockets
 
         public void Abort()
         {
-            _abortSource.Cancel();
+            _state = WebSocketState.Aborted;
             WebSocket?.Abort();
         }
 
-        public async Task ConnectAsync(Uri uri, CancellationToken cancellationToken, ClientWebSocketOptions options)
+        public Task ConnectAsync(Uri uri, CancellationToken cancellationToken, ClientWebSocketOptions options)
         {
-            try
-            {
-                cancellationToken.ThrowIfCancellationRequested();  // avoid allocating a WebSocket object if cancellation was requested before connect
-                CancellationTokenSource? linkedCancellation;
-                CancellationTokenSource externalAndAbortCancellation;
-                if (cancellationToken.CanBeCanceled) // avoid allocating linked source if external token is not cancelable
-                {
-                    linkedCancellation =
-                        externalAndAbortCancellation =
-                        CancellationTokenSource.CreateLinkedTokenSource(cancellationToken, _abortSource.Token);
-                }
-                else
-                {
-                    linkedCancellation = null;
-                    externalAndAbortCancellation = _abortSource;
-                }
+            cancellationToken.ThrowIfCancellationRequested();
 
-                using (linkedCancellation)
-                {
-                    WebSocket = new BrowserWebSocket();
-                    await ((BrowserWebSocket)WebSocket).ConnectAsyncJavaScript(uri, externalAndAbortCancellation.Token, options.RequestedSubProtocols).ConfigureAwait(continueOnCapturedContext: true);
-                    externalAndAbortCancellation.Token.ThrowIfCancellationRequested();
-                }
-            }
-            catch (Exception exc)
-            {
-                if (_state < WebSocketState.Closed)
-                {
-                    _state = WebSocketState.Closed;
-                }
-
-                Abort();
-
-                switch (exc) {
-                    case WebSocketException:
-                    case OperationCanceledException _ when cancellationToken.IsCancellationRequested:
-                        throw;
-                    default:
-                        throw new WebSocketException(WebSocketError.Faulted, SR.net_webstatus_ConnectFailure, exc);
-                }
-            }
+            var ws = new BrowserWebSocket();
+            WebSocket = ws;
+            return ws.ConnectAsync(uri, options.RequestedSubProtocols, cancellationToken);
         }
     }
 }