… THIRD-PARTY-NOTICES.TXT
… src/libraries/System.Net.WebSockets.Client/src/System/Net/WebSockets/BrowserWebSockets/BrowserWebSocket.cs
… src/libraries/System.Net.WebSockets.Client/src/System/Net/WebSockets/BrowserWebSockets/ReceivePayload.cs
… src/libraries/System.Net.WebSockets.Client/tests/SendReceiveTest.cs
… src/mono/wasm/runtime/binding_support.js