- Eliminate as many copies of the data as possible, copy data synchronously when possible
- move buffers to js side
- move most of the implementation to js
- when send buffer became larger than 64k, resolve "send" Task only after data departed to OS via `bufferedAmount === 0`
- resolve "close" Task only after we received "close" response from browser
- introduce simple cancelable promise on JS side
- Abort() cancels all the pending send/receive/open/close Tasks. 
- Abort() sends "close" to server, but doesn't wait for response
- The cancelationToken cancels only the current operation and sets the whole WS into Aborted state. Other pending operations would finish normally but `BrowserWebSocket` would throw on any subsequent operations, except "close".

The previous `BrowserWebSocket` implementation did following actions, which the proposed implementation doesn't do 
- `SendAsync` didn't block task on `endOfMessage` for large pending transmission
- `CloseOutputAsync` self produced onClose message for it's receivers
- `ConnectAsync` sent "close" message to server and self produced onClose message for it's receivers, when canceled
- `ReceiveAsync` sent "close" message to server , when canceled

Fixes https://github.com/dotnet/runtime/issues/44614
Fixes https://github.com/dotnet/runtime/issues/58202
Makes https://github.com/dotnet/runtime/issues/57519 to just report slow test to console instead of fail.
Makes https://github.com/dotnet/runtime/issues/53957 to just report slow test to console instead of fail.
Fixes https://github.com/dotnet/runtime/issues/58031