@pavelsavara 2021-08-30
Right now it doesn't handle all the abort/cancelation/dispose scenarios. Work in progress ... stay tuned :-D