@pavelsavara 2021-09-01
I tested with Chrome, Edge and Firefox locally.