@pavelsavara 2021-09-02
> Why does this PR remove Add/RemoveEventListener?

Because we don't need it anymore. Do you think that it should be part of the public API going forward ?