@pavelsavara 2021-09-02
The implementation in the PR right now is running unit test suite 5x slower than main branch. It's caused by blocking send operation until the buffer is empty.

We had chat with @kg and here are conclusions
- we would not block send operation's Task if the `bufferedAmount` is under 64k. This would improve throughput close to previous implementation and also apply backpressure for apps sending large data.
- we would keep the Add/RemoveEventListener code
- we would not automatically send 'close' to the server on operation cancelation. But we should allow the user to use CloseAsync after any cancelation.