@pavelsavara 2021-09-06
I tested wall clock performance while running unit test suite.
- Old code: 160 seconds
- New code: 157 seconds without waiting for "close" and "send"
- New code: 177 seconds with waiting for "close" and "send"

My conclusion is that paying with wall clock time for consistency is fine.