@pavelsavara 2021-09-07
[Microbenchmark of the work with buffers](https://github.com/dotnet/runtime/pull/58730)
| measurement | old time |new time|
|-:|-:|-:|
|        WebSocket, PartialSend_1B |     0.0029ms |0.0315ms |
|      WebSocket, PartialSend_64KB |     0.2612ms |2.3403ms |
|       WebSocket, PartialSend_1MB |     4.2909ms |2.1455ms |
|     WebSocket, PartialReceive_1B |     0.0086ms |0.0174ms |
|   WebSocket, PartialReceive_10KB |     0.0001ms |0.0909ms |

It shows that managed memcpy is cheap compared to crossing the JS boundary with each of many small buffers.
On the other hand, the old implementation is easy to run into OutOfMemoryException.

Right now I think that we need to do buffering inside of a message on managed side and buffering of whole messages on the JS side. We need to discuss if it makes sense to also push split extra large messages and keep data on the JS side for them. How many bytes should be the threshold ?