@pavelsavara 2021-09-15
Notes I made during code walkthrough with @CarnaViire 

	- OutOfMemory exception, it should not be translated as WebSocketException
	- order of response ints[] on receive, name the index constants
	- rename promise_control for open
	- ws[BINDING.wasm_ws_pending_receive_promise_queue].drain -> int[]
	- fixed buffer for middle of array, use ArraySegment and offset instead
	- _mono_wasm_web_socket_send_buffering message_ptr-> message_root
	- utf8 test outerloop on coreCLR
	- bufferedAmount: "This value does not reset to zero when the connection is closed" what about CloseSent ?
	- assert: when we get message and the receivers queue is not empty, there need to be empty even queue, 
    - assert: mirror of above in the receiver method
	- code 1001/1005/1006 on abort https://developer.mozilla.org/en-US/docs/Web/API/CloseEvent/code
	- SendReceive_VaryingLengthBuffers_Success - no change for CoreCLR, keep the active issue for why it's so slow