@kg 2021-09-16
> 
> 
> Follow up on yesterday conversation with @CarnaViire about CloseOutput from client side.
> There is no such API on browser WS.
> And therefore we call `WebSocket.close()` instead.
> Chrome immediately issues 'close' event and it would not call and any 'message' callbacks after that, even if there were some messages in flight from the server.
> 
> So, we are losing the messages from server after `CloseOutputAsync`, as compared to managed/desktop implementation which delivers them.
> 
> The alternative is to not call `WebSocket.close()` when `CloseOutputAsync`, but that seems even worse.

Hm, this is a huge failure on browser vendors' part. I don't like the data loss here. Is there some way we could signal this behavior difference to the developer?