@pavelsavara 2021-09-16
The other differences in behavior are 
- we are not _streaming_ partial messages, we delay them till `endOfMessage`.
- we are sending `close(1000)` NormalClosure on `Abort()`