@pavelsavara 2021-09-16
> Is there some way we could signal this behavior difference to the developer?

We agreed with @kg to issue warning into console once per app lifetime. 
`WARNING: Web browsers do not support closing the output side of a WebSocket. CloseOutputAsync has closed the socket and discarded any incoming messages.`